+++ src/client/components/Editors/Dialect/index.tsx
@@ -6,7 +6,6 @@ import * as types from '../../../types'
 import Layout from './Layout'
 
 export interface DialectProps {
-  type?: string
   format?: string
   dialect?: types.IDialect
   externalMenu?: { section: string }
