+++ src/client/components/Editors/Resource/Layout.tsx
@@ -88,7 +88,6 @@ function LayoutWithMenu() {
           <Box>
             <Box hidden={!section.startsWith('dialect')}>
               <Dialect
-                type={type}
                 format={format}
                 dialect={dialect}
                 externalMenu={externalMenu}
